🚚 move(#21): mdx
