+++ src/lib/mdx/index.ts
@@ -1,0 +1,1 @@
+export { mdx2html } from "./mdx2html";
