+++ src/pages/articles/id.client.test.tsx
@@ -2,8 +2,8 @@ import { mockArticles, mockCategories, mockConfig, mockPickup, mockPopularArticl
 import { render, screen } from "jest/test-utils";
 import { server } from "mocks/msw/server";
 
+import { mdx2html } from "@/lib/mdx";
 import { formatPageTitle } from "@/utils/formatter";
-import { mdx2html } from "@/utils/mdx/mdx2html";
 
 import ArticleDetail from "./[id].page";
