+++ src/pages/preview/[id].page.tsx
@@ -2,13 +2,13 @@ import type { GetServerSideProps } from "next";
 import type { Params } from "next/dist/shared/lib/router/utils/route-matcher";
 
 import { HtmlHeadNoIndex } from "@/components/functions/meta";
+import { mdx2html } from "@/lib/mdx";
 import type { ArticleDetailProps, ArticlesStaticProps } from "@/pages/articles/[id].page";
 import ArticleDetail from "@/pages/articles/[id].page";
 import { getPickupArticles, getPopularArticles } from "@/services/article";
 import type { TArticle } from "@/types";
 import { isDraft } from "@/utils/article";
 import { fetchArticle, fetchCategories, fetchConfig } from "@/utils/fetcher";
-import { mdx2html } from "@/utils/mdx/mdx2html";
 
 const ArticlePreview = (props: ArticleDetailProps) => {
   return (
